requirements.txt
gradio==3.16.2
transformers==4.26.1
albumentations==1.3.0
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
pytorch-lightning==1.7.7
omegaconf==2.2.3
einops==0.4.1
open_clip_torch==2.7.0
timm==0.9.2
safetensors==0.3.0
accelerate==0.17.1
clean-fid==0.1.29
imageio==2.27.0
joblib==1.2.0
llvmlite==0.39.1
numba==0.57.1
pillow==9.4.0
protobuf==3.20.0
xformers==0.0.17
torchmetrics==0.11.4
torch==2.0.0
numpy==1.22.3